+++ NEWS.md
@@ -1,3 +1,8 @@
+# phstemplates 1.1.1
+
+* Added Posit Workbench options and quit() to script template.
+* New addin to update the metadata in existing R scripts for latest update date and R version run on.
+
 # phstemplates 1.1.0
 
 * Added PHS shiny app template, accessible through the RStudio menus or through phsshinyapp().